requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.28.2
python-dotenv==1.0.0
openai==0.28.0